Move CURLHandle to separate header
