+++ extension/httpfs/httpfs_curl_client.cpp
@@ -7,6 +7,10 @@
 #include <sys/stat.h>
 #include "duckdb/common/exception/http_exception.hpp"
 
+#ifndef EMSCRIPTEN
+#include "httpfs_curl_client.hpp"
+#endif
+
 namespace duckdb {
 
 // we statically compile in libcurl, which means the cert file location of the build machine is the
